# dwes06/dwes06Servicio/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Talleres y Ubicaciones",
        "description": "Listado de URI's de la API de Talleres y Ubicaciones",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
        "/api/ubicaciones/{idUbicacion}/creartaller": {
            "post": {
                "tags": [
                    "Talleres"
                ],
                "summary": "Almacena un taller en la base de datos.",
                "operationId": "374717333be8337e8dd7975990823950",
                "parameters": [
                    {
                        "name": "idUbicacion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del taller introducidos en el formulario",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Taller de lectura"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Taller de lectura en la biblioteca"
                                    },
                                    "dia_semana": {
                                        "type": "string",
                                        "example": "L"
                                    },
                                    "hora_inicio": {
                                        "type": "string",
                                        "example": "10:00"
                                    },
                                    "hora_fin": {
                                        "type": "string",
                                        "example": "11:00"
                                    },
                                    "cupo_maximo": {
                                        "type": "integer",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "resultado": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "datos": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "ubicacion_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Taller de lectura"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Taller de lectura en la biblioteca"
                                                    },
                                                    "dia_semana": {
                                                        "type": "string",
                                                        "example": "L"
                                                    },
                                                    "hora_inicio": {
                                                        "type": "string",
                                                        "example": "10:00:00"
                                                    },
                                                    "hora_fin": {
                                                        "type": "string",
                                                        "example": "11:00:00"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "La ubicación no existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errores": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "nombre": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "El nombre es obligatorio."
                                                        }
                                                    },
                                                    "dia_semana": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "El día de la semana es obligatorio."
                                                        }
                                                    },
                                                    "hora_inicio": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "La hora de inicio es obligatoria."
                                                        }
                                                    },
                                                    "hora_fin": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "La hora de fin es obligatoria."
                                                        }
                                                    },
                                                    "cupo_maximo": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "El cupo máximo es obligatorio."
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/talleres/{idTaller}": {
            "delete": {
                "tags": [
                    "Talleres"
                ],
                "summary": "Elimina un taller con un determinado id de la base de datos.",
                "operationId": "bbf4303f2f4e81d2b16f27bcafc6775d",
                "parameters": [
                    {
                        "name": "idTaller",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "resultado": {
                                            "type": "string",
                                            "example": "eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "resultado": {
                                            "type": "string",
                                            "example": "No existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}